chore: move prune test to its own file
